extracted cacher from gmsh stuff
